refactor: move total installs to downloads
